+++ Samples/Packages/com.unity.entities/Unity.Entities.Hybrid.Tests/ComponentGroupTransformAccessArrayTests.cs
@@ -38,7 +38,7 @@ namespace Unity.Entities.Tests
 	    [Test]
 		public void EmptyTransformAccessArrayWorks()
 	    {
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 			Assert.AreEqual(0, ta.length);
 	        group.Dispose();
@@ -50,10 +50,10 @@ namespace Unity.Entities.Tests
 	        go.AddComponent<TransformAccessArrayTestTagComponent>();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go.GetComponent<GameObjectEntity>().OnEnable();
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(1, ta.length);
-	        group.Dispose();
+
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go.GetComponent<GameObjectEntity>().OnDisable();
 	        Object.DestroyImmediate(go);
@@ -65,7 +65,7 @@ namespace Unity.Entities.Tests
 	        go.AddComponent<TransformAccessArrayTestTagComponent>();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go.GetComponent<GameObjectEntity>().OnEnable();
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(1, ta.length);
 
@@ -76,7 +76,6 @@ namespace Unity.Entities.Tests
 	        ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(2, ta.length);
 
-	        group.Dispose();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go.GetComponent<GameObjectEntity>().OnDisable();
 	        go2.GetComponent<GameObjectEntity>().OnDisable();
@@ -91,7 +90,7 @@ namespace Unity.Entities.Tests
 	        go.AddComponent<TransformAccessArrayTestTagComponent>();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go.GetComponent<GameObjectEntity>().OnEnable();
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(1, ta.length);
 
@@ -120,7 +119,7 @@ namespace Unity.Entities.Tests
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go2.GetComponent<GameObjectEntity>().OnEnable();
 
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(2, ta.length);
 
@@ -131,7 +130,6 @@ namespace Unity.Entities.Tests
 	        ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(1, ta.length);
 
-	        group.Dispose();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go2.GetComponent<GameObjectEntity>().OnDisable();
 	        Object.DestroyImmediate(go2);
@@ -149,7 +147,7 @@ namespace Unity.Entities.Tests
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go2.GetComponent<GameObjectEntity>().OnEnable();
 
-	        var group = m_Manager.CreateComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
+	        var group = EmptySystem.GetComponentGroup(typeof(Transform), typeof(TransformAccessArrayTestTag));
 	        var ta = group.GetTransformAccessArray();
 	        Assert.AreEqual(2, ta.length);
 
@@ -159,7 +157,6 @@ namespace Unity.Entities.Tests
 
 	        Assert.AreEqual(2, ta.length);
 
-	        group.Dispose();
 	        // Execute in edit mode is not enabled so this has to be called manually right now
 	        go2.GetComponent<GameObjectEntity>().OnDisable();
 	        Object.DestroyImmediate(go2);