+++ ReleaseNotes.md
@@ -1,6 +1,7 @@
 # 0.0.5
 ## Changes
 * Throw ArgumentException when creating an entity with component data exceeding chunk size (64kb)
+* EntityManager.CreateComponentGroup is no longer public, use ComponentSystem.GetComponentGroup instead
 
 # 0.0.4